src/pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "openenv-core"
version = "0.1.1"
description = "Core components and cli for OpenEnv - HTTP-based agentic environments"
readme = "./core/README.md"
requires-python = ">=3.10"
license = {text = "BSD-3-Clause"}
authors = [
    {name = "Meta Platforms, Inc.", email = "opensource@meta.com"}
]
keywords = ["environment", "agent", "http", "docker", "fastapi"]

dependencies = [
    "fastapi>=0.104.0",
    "pydantic>=2.0.0",
    "uvicorn[standard]>=0.24.0",
    "requests>=2.25.0",
    "typer>=0.12.0",
    "pyyaml>=6.0.0",
    "huggingface_hub>=0.24.0",
    "rich>=13.0.0",
    "tomli>=2.0.1",
    "tomli-w>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]

# Fleet runtime integration (optional)
fleet = [
    "mcp>=1.0.0",
    "fleet-sdk>=0.2.79",
]

[project.scripts]
openenv = "openenv_cli.__main__:main"

[project.urls]
Homepage = "https://github.com/facebookresearch/OpenEnv"
Repository = "https://github.com/facebookresearch/OpenEnv"
Documentation = "https://github.com/facebookresearch/OpenEnv/blob/main/README.md"
"Bug Tracker" = "https://github.com/facebookresearch/OpenEnv/issues"

[tool.setuptools]
package-dir = { "openenv_core" = "core", "openenv_cli" = "openenv_cli"}
packages = [
    "openenv_core",
    "openenv_core.containers",
    "openenv_core.containers.runtime",
    "openenv_core.env_server",
    "openenv_core.tools",
    "openenv_cli",
    "openenv_cli.commands",
    "openenv_cli.templates"
]

[tool.setuptools.package-data]
"openenv_cli" = ["templates/**/*"]
